Add initialisation to new clock drivers.
Both the IIgs and NoSlotClock drivers now have initialisation that looks
for valid clock hardware before linking in the code that reads it.

The system startup program now checks for a valid clock before updating
the system, and is now safe to use across systems with different clock
hardware.

The NoSlotClock version is based on the reverse engineered ProDOS
install found at:
  https://gist.github.com/mgcaret/ae2860c754fd029d2640107c4fe0bffd